+++ src/globals.d.ts
@@ -4,6 +4,7 @@ import type { UserAnswersDb } from "./composables/db";
 declare global {
     interface Window {
         idb: IDBPDatabase<UserAnswersDb>,
-        getDb: () => Promise<IDBPDatabase<UserAnswersDb>>
+        getDb: () => Promise<IDBPDatabase<UserAnswersDb>>,
+        getPathName: () => string;
     }
 }